Connections, no matches, and email checking

+++ src/server.js
@@ -15,6 +15,9 @@ import Skill from './models/skill_model';
 
 dotenv.config({ silent: true });
 
+// Make this true to reset the DB on startup
+const resetDB = true;
+
 const setUpDB = () => {
   const self = new User();
   self.firstName = 'Pass';
@@ -209,14 +212,16 @@ const setUpDB = () => {
 // DB Setup
 const mongoURI = process.env.MONGODB_URI || 'mongodb://localhost/passon';
 mongoose.connect(mongoURI, () => { // this code clears the database on every server run
-  mongoose.connection.db.dropDatabase(() => {
-    mongoose.connection.close(() => {
-      mongoose.connect(mongoURI);
-      console.log('reconnected');
-      setUpDB();
-      console.log('default db set up');
+  if (resetDB) {
+    mongoose.connection.db.dropDatabase(() => {
+      mongoose.connection.close(() => {
+        mongoose.connect(mongoURI);
+        console.log('reconnected');
+        setUpDB();
+        console.log('default db set up');
+      });
     });
-  });
+  }
 });
 // set mongoose promises to es6 default
 mongoose.Promise = global.Promise;
